Version 1.0 is now out of beta!
git-svn-id: http://taiga.googlecode.com/svn/trunk@282 77add795-4bdf-f0f6-a574-f4d068491005

+++ update.h
@@ -40,7 +40,7 @@ public:
   void SetDownloadPath(const wstring& path);
 
   const GenericFeedItem* FindItem(const wstring& guid) const;
-  unsigned long GetVersionValue(int major, int minor, int revision) const;
+  unsigned long long GetVersionValue(int major, int minor, int revision) const;
 
   HttpClient client;
   vector<GenericFeedItem> items;
